fix(common): use the OpenAPI component list to place the public boundary
The `/internal` split used "no open-source-side package imports it" as the test
for whether a type is public. That test is wrong in one direction, and it cost
17 symbols.

A type can be a published wire contract that no TypeScript consumer ever names,
because the generated Java, Go and Python clients are generated from the OpenAPI
document rather than from the SDK. `sts-token-types.ts` (10 published components)
and `store/conversation-state.ts` (7, including `ConversationState` and
`TextArtifactReference`, which other components `$ref`) are exactly that. Moving
them behind a subpath silently narrowed the published API surface, and every
TypeScript build in the monorepo stayed green while it happened.

Both modules go back to the package root, and `internal/no-published-components.test.ts`
now asserts that nothing reachable from `src/internal/index.ts` appears in
the published OpenAPI document, so the next move of this kind fails a
test instead of shipping.

That correction also removes one of the three edges that made the remaining
internal modules non-extractable — `interaction.ts` -> `conversation-state.ts`
was public-to-public all along. The other two were misplacements rather than
coupling, and are fixed here:

- `validateProjectSearchPropertyMappings` was a server-side validator living in
  the public types module, which is why `project.ts` had to reach into
  `view-validation-helpers.js`. It moves to `internal/project-validation.ts`
  along with the three private constants it owns; its tests move with it.
- `ParentClosePolicyType` is part of the public workflow DSL — clients author
  `parentClosePolicy` on a child-workflow step — so the union moves to
  `store/dsl-workflow.ts`. Only the `ParentClosePolicy` runtime constant, used by
  the platform's own workflow-authoring code, stays internal.

No public module imports from the internal tree any more, so the 14 modules
behind `/internal` are now free of inbound edges from the public tier.
