+++ packages/common/src/index.ts
@@ -71,6 +71,7 @@ export type {
     ZenoBulkContentObjectExportSplitShardResponse,
 } from './store/store.js';
 export * from './sts-errors.js';
+export * from './sts-token-types.js';
 export * from './tenant.js';
 export * from './tool-execution.js';
 export * from './tools.js';
