+++ packages/common/src/store/index.ts
@@ -3,6 +3,7 @@ export * from './agent-approval.js';
 export * from './agent-run.js';
 export * from './collections.js';
 export * from './common.js';
+export * from './conversation-state.js';
 export * from './doc-analyzer.js';
 export * from './dsl-workflow.js';
 export * from './grounded-extraction.js';
